requirements.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
notion-client==0.4.0
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
